requirements.txt
Flask
mdurl
pydantic
linkify-it-py
